requirements.txt
Dependencies:

shellescape
Naked
PyYAML
pycryptodome
Tkinter

